Summary: Create skeleton for the perceptron server
A Neural Network is made up of many connected Perceptrons.  I have a pretty good idea about how perceptrons work from the tutorial at http://www.trapexit.org/Erlang_and_Neural_Networks. So it makes sense to start on this.

Perceptrons are nodes in the network that are passed a value that they then pass on through their output connections.  Perceptrons maintain state information on their inputs, outputs, and weight vectors and have to wait on and respond to messages.  For this, the gen_server behaviour is well suited.  I have added a generic skeleton for the perceptron gen_server.  The actual logic will be added shortly

Please note that I didn't use the gen_server:start_link with the optional server name argument.  I need a ton of perceptrons, so they will be referred to by their PIDs.