Summary: Create methods to hook together perceptrons
Neural nets are networks of connected perceptrons.  They can be very
naturally represented by directed graphs.  These weighted connections
are the backbone of the contraption.

My first priority for creating a server representation is thus to
create a mechanism for connecting perceptrons to each other.  This
will be done through gen_server casts.

gen_server has two methods for communicating with the server, casts
and calls.  Casts are asynchronous and do not return a reply.  Calls
are synchronous and return a reply.  Calls basically act like normal
function calls, except against a server.

For connecting perceptrons together, I am using a cast because the
only possible return value would be an 'ok' confirmation.  Connecting
perceptrons should always work unless something goes horribly wrong,
in which case throwing an exception, failing, and burning that mother
down are the appropriate responses.

The input and output connections are the beginning of state for our
server, so we must update the state record to represent this.  I give
inputs and outputs a default value of an empty list.  Each element in
either of those lists represents a weighted connection, so I further
define a connection record which consists of the PID address of the
connected node and the weight.  All connections are started at the
weight of 0.5.

The actual magic happens in the handle_cast callbacks.  I define
function clauses that match the messages {connect_to_input, PID} and
{connect_to_output, PID}.  These then extract either inputs or outputs
from the server state, copy the state record, changing the appropriate
list to contain the new connection.

The final step for this commit is to write a connect API method that
abstracts this tomfoolery from the user. It takes two nodes, A and B,
and connects B as an output to A and A as an input to B.