Summary: Implement basic feed-forward messages
The actual operation of a feed-forward neural network is fairly
simple.  A perceptron is stimulated by the output values of its input
nodes, the dot product of the weights and input values is taken, the
dot product is adjusted by a thresholding function (in this case a
simple sigmoid), and is then passed on as stimulation to the
perceptron's outputs.

To implement this, I first have to modify the connections record to
contain a field for the value emitted by the node.  This is only
relevant for input connections, but maintaining a default value (0)
will not affect output connections. The stimulation is handled through
the {stimulate, From, Value} cast.  The inputs are updated with the
new value (update_inputs/3).  The output value is then calculated
(output_value/1).  The output value is then passed on to the output
connections (stimulate_outputs/2).  If the node is terminal, then the
output is written to the shell (for now).

I also add a stimulate method to the API.  This is actually used in
stimulate_outputs to handle the actual message passing.

You might notice that this causes a recomputation of the output value
and restimulation of all connection values whenever a single input is
updated.  The proper behavior is to wait for all inputs to be updated.
This will be handled in a later installment.

Hopefully the code is fairly understandable at this stage.  I did not
mention all of the details of how everything works, but I hope that a
reading of the code will yield understanding.