Summary: Create entry point for the neural network
The perceptrons can now do most of what they need in the forward
direction.  The main thing left is to define an entry point for the
network.  As it currently stands, a perceptron with no inputs cannot
be stimulated in a way that will cause the proper value to be passed
on.  If you try perceptron:stimulate(Node1, 0.5O) at the shell it will
cause output to be produced, but it will have simply passed on the
dot product of the empty list, 0.

To fix this, I add a cast of {pass, Value} to the server.  This simply
takes a value, thresholds it, and stimulates the output connections.
It does not modify state.  A pass/2 method is then added to the API.

Now that the neural net is somewhat functional, it is a good time to
fix that messy issue of recalculating everytime a single input is
changed.  This is corrected by maintaining a list of stale inputs in
the server state.  This list is only composed of node PIDs.
update_inputs is modified to adjust the list of stale inputs on
stimulation.  When an input stimulates the perceptron, it is no longer
considered stale.  After the outputs are stimulated, all of the inputs
are marked stale.

You can now experiment in the shell with this.

4> {ok, A} = perceptron:start_link().
{ok,<0.42.0>}
5> {ok, B} = perceptron:start_link().
{ok,<0.44.0>}
6> perceptron:connect(A,B).
ok
7> perceptron:pass(A, 0.5).
<0.44.0>: 0.5771853801446523
ok
8> perceptron:pass(A, 0.9).
<0.44.0>: 0.5879445592675961
ok
9> perceptron:pass(A, 45).
<0.44.0>: 0.6224593312018546
ok

This shell session just created a simple two node network and fed it
values.