Summary: Create a server to manage the entire network
As things currently stand, I can create a neural network, pass it
input, and read the output off of the stdout.  This is all well and
good, but it makes using the output values in a program basically
impossible.  I want to be able to call a function and have the return
value be the output of the network.

To accomplish this, a new gen_server, neural_network, is made.  This
requires a modification to the message format of the perceptrons.  I
now need to incorporate a Requester parameter into the tuple.  This is
passed through so that the terminal node can send the output as a
message to the Requester process.  This parameter shall occur
immediately following the cast name atom.

I start from the standard gen_server skeleton for neural_network.
Notice this time that start_link registers this server under the
module name.

This server requires a network to exist for it to function.  I define
a simple A -> B network in init/1.  Eventually this will be replaced
with code to construct a network from a tree structure.  The server
state simply consists of the entry node to the network.  Currently
only one input node is allowed, this will be changed in time.

The important part of this is the {calculate, Value} call.  This
passes the value to the start node of the network and then waits for a
message from the terminal node of the network.  The calculated value
is the return value of the call.  This is encapsulated by the
calculate/1 API function.  Notice that it is given an infinite timeout
value.  This is done to ensure that the server blocks until the
network does its thing and responds.

This can be executed from the prompt.  Notice that the calculated
values are the same as were found from the identical network in the
last commit.

25> neural_network:start_link().
{ok,<0.99.0>}
26> neural_network:calculate(0.5).
0.5771853801446523