Summary: Implement back propagation learning mode
Being able to calculate values by feeding through a neural network is
fantastic, but it, like a film critic screening a Vin Diesel movie, is
just missing the point.  A neural network without a learning mode is
like a sad panda tasked to roam the earth for all eternity.  The
simplest method for our simple neural network is called back
propagation.  Essentially it consists of subtracting the output from
the expected value to form a delta and then propagating that back
through the network.  Just as output values were found by taking the
dot product of the inputs, going in the reverse direction involves
calculating the deltas from the dot product of the outputs.  The
weights are then adjusted through a learning function.

First, I need to change the state to add stale_outputs.  I add a
function update_outputs to keep the deltas and weights synced up.
update_outputs looks almost the same as update_inputs except with the
addition of weight, one would conclude that those functions can be
generalized and sacrifices to the DRY gods could proceed as normal.
The problem arises due to the use of records, and the bad habit of
records to be preprocessor trickery.  Much repetition will probably
come of this, unfortunately.

The next step is handling the {learn, From, Weight, Delta} cast.  The
body of this handle_cast is very similar to the one that handles
stimulation. This logically follows because the process is basically
the same but in the opposite direction.  The major difference is that
the return point for this function clause is in the case statement.
The return values are sufficiently different in each case to warrant
this change.

The function backpropagate_inputs works just like stimulate_outputs,
but in reverse.  The delta_value function is just the dot product of
the weights and the values from the output list.  Just as the
stimulate function encapsulated the server cast behind stimulation,
backpropagate encapsulates the learn server cast.

The real meat of learning is in update_input_weight.  The algorithm
for learning for adjusting the weights is (feel free to throw this in
your TeX machine)

$$ w'_x=w_x + \eta \delta \frac{df(e)}{de} y_x $$

where w_x is the weight coming in from input x, y_x is the value of
the input, delta is the value determined by delta_value, eta is the
learning rate (set to 0.5 by fiat here), and df(e)/de is the output
value calculated using the derivative of the sigmoid function instead
of the sigmoid function. The multiplication of eta, delta, and the
output derivative creates a coefficient that is multiplied by a
specific input's value and then added to that input's weight to yield
the new weight.

A better explanation of the process is found at http://galaxy.agh.edu.pl/~vlsi/AI/backp_t_en/backprop.html